// src/app/component/enums/MessageEnum.ts
export enum MessageEnum {
  CONNECTION_REFUSED = "connection.refused",
  CONNECTION_REFUSED_DETAIL = "connection.refused.detail",

  CONFIRM_SUCCESS = 'confirm.success',
  CONFIRM_REJECT = 'confirm.reject',
  CONFIRM_WARNING = 'confirm.warning',
  CONFIRM_INFO = 'confirm.info',

  CONFIRM_SUCCESS_200 = 'confirm.success.200',
  CONFIRM_SUCCESS_204 = 'confirm.success.204',
  CONFIRM_REJECT_400 = 'confirm.reject.400',
  CONFIRM_REJECT_404 = 'confirm.reject.404',
  CONFIRM_REJECT_500 = 'confirm.reject.500'
}